+++ src/main/java/org/example/backend/common/weight/controller/WeightController.java
@@ -1,10 +1,10 @@
 package org.example.backend.common.weight.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.example.backend.baseball.weight.KboTeamWeight;
-import org.example.backend.baseball.weight.UserKboWeight;
-import org.example.backend.baseball.weight.KboTeamWeightRepository;
-import org.example.backend.baseball.weight.KboWeightService;
+import org.example.backend.common.weight.entity.KboTeamWeight;
+import org.example.backend.common.weight.entity.UserKboWeight;
+import org.example.backend.common.weight.repository.KboTeamWeightRepository;
+import org.example.backend.common.weight.service.WeightService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -16,14 +16,15 @@ import java.util.Map;
 @RequiredArgsConstructor
 public class WeightController {
 
-    private final KboWeightService kboWeightService;
+    private final WeightService weightService;
     private final KboTeamWeightRepository kboTeamWeightRepository;
 
-    @GetMapping("/recommend/kbo")
-    public ResponseEntity<List<Map.Entry<String, Double>>> recommendKboTeams(@RequestBody UserKboWeight userKboWeight) {
+    @PostMapping("/recommend")
+    public ResponseEntity<List<Map.Entry<String, Double>>> recommendTeams(@RequestBody UserKboWeight userKboWeight) {
+
 
         List<KboTeamWeight> teamWeights = kboTeamWeightRepository.findAll();
-        List<Map.Entry<String, Double>> result = kboWeightService.kboRankTeams(teamWeights, userKboWeight);
+        List<Map.Entry<String, Double>> result = weightService.kboRankTeams(teamWeights, userKboWeight);
 
         return ResponseEntity.ok(result);
     }
