+++ src/main/java/org/example/backend/common/weight/repository/KboTeamWeightRepository.java
@@ -1,5 +1,6 @@
-package org.example.backend.baseball.weight;
+package org.example.backend.common.weight.repository;
 
+import org.example.backend.common.weight.entity.KboTeamWeight;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface KboTeamWeightRepository extends JpaRepository<KboTeamWeight, Long> {
